[FIX] #232 불필요한 Fragment 제거
